Implemented player interaction - John Francis
Player can now take damage from spikes and water and press buttons that open up barriers for each floor.
NOTE:
Damage does not actually do anything, the node just runs whenever the player is made to take damage. No animation plays as of yet.
Water texture does not animate and I did not add a background.
New obstacles do not delete within a distance of the player.
Most of the code is spaghetti code. I will need to reorganize and possibly optimize a lot of it.
Lots of other unfinished stuff as well on my part.